move sync service to event service
