+++ idl/event.proto
@@ -2,6 +2,7 @@ syntax = "proto3";
 package limit.event;
 
 import "auth.proto";
+import "event.types.proto";
 
 message Event {
     // uuid of the event
@@ -14,21 +15,8 @@ message Event {
     string sender = 3;
     // event data
     oneof detail {
-        Message message = 4;
+        limit.event.types.Message message = 4;
     }
-}
-
-message Message {
-    // TODO: it should be subscription
-    // receiver uuid of the message
-    string receiver_id = 4;
-    // TODO: no need to specify receiver_server
-    string receiver_server = 5;
-    // text should be encrypted
-    // if using v1 or v2 version text could be empty
-    string text = 6;
-
-    map<string, string> extensions = 7;
 }
 
 message ReceiveEventsRequest {
@@ -48,9 +36,34 @@ message SendEventResponse {
     string event_id = 1;
 }
 
+message SynchronizeRequest {
+    // auth
+    limit.auth.Auth token = 1;
+    // target subscriber
+    string subscription = 2;
+    oneof starting_point {
+        // synchronization start timestamp
+        // the default is now
+        int64 timestamp = 3;
+        // the ID of the event to start synchronization
+        string event_id = 4;
+    }
+    // offset for synchronization
+    // the default is +50 (49 events backwards from the starting point)
+    sint32 offset = 5;
+    // returns all type of events if not specified
+    limit.event.types.Types filter = 6;
+}
+
+message SynchronizeResponse {
+    repeated limit.event.Event events = 1;
+}
+
 service EventService {
-   // receiving all events from the server
-   rpc ReceiveEvents (ReceiveEventsRequest) returns (stream Event);
-   // sending event to the server
-   rpc SendEvent (SendEventRequest) returns (SendEventResponse);
+    // receiving all events from the server
+    rpc ReceiveEvents (ReceiveEventsRequest) returns (stream Event);
+    // sending event to the server
+    rpc SendEvent (SendEventRequest) returns (SendEventResponse);
+    // synchronize from server
+    rpc Synchronize(SynchronizeRequest) returns (SynchronizeResponse);
 }
