+++ idl/event.types.proto
@@ -1,0 +1,20 @@
+syntax = "proto3";
+package limit.event.types;
+
+enum Types {
+    ALL = 0;
+    MESSAGE = 1;
+}
+
+message Message {
+    // TODO: it should be subscription
+    // receiver uuid of the message
+    string receiver_id = 4;
+    // TODO: no need to specify receiver_server
+    string receiver_server = 5;
+    // text should be encrypted
+    // if using v1 or v2 version text could be empty
+    string text = 6;
+
+    map<string, string> extensions = 7;
+}
