--- limit-server-event-sync/Cargo.toml
@@ -1,56 +1,0 @@
-[package]
-name = "limit-server-message-sync"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-anyhow = "1"
-async-trait = "0.1"
-tokio = { version = "1", features = ["full"] }
-chrono = { version = "0.4", features = ["serde"] }
-jsonwebtoken = "8.1"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-
-pilota = "*"      # we recommend to use the latest framework version for new features and bug fixes
-volo = "*"        # we recommend to use the latest framework version for new features and bug fixes
-volo-grpc = "*"   # we recommend to use the latest framework version for new features and bug fixes
-volo-gen = { path = "../volo-gen" }
-rand = "*"
-tracing = "0.1"
-futures = "*"
-async-channel = "*"
-dashmap = "*"
-tokio-util = "*"
-
-
-limit-server-auth = { path = "../limit-server-auth" }
-limit-db = { path = "../limit-db" }
-limit-am = { path = "../limit-am" }
-limit-config = {path = "../limit-config"}
-limit-utils = {path = "../limit-utils"}
-
-[dependencies.redis]
-version = "*"
-features = [
-    "aio",
-    "cluster",
-    "r2d2",
-    "tokio-comp"
-]
-
-[dependencies.diesel]
-version = "2.0.0"
-features = ["sqlite", "uuid", "chrono", "r2d2"]
-
-[dependencies.uuid]
-version = "1.2.1"
-features = [
-    "v4",
-    # "v7",
-    "fast-rng",
-    "macro-diagnostics",
-    "serde",
-]
